Renamed module (DH-17600)
