+++ src/crossModule/index.ts
@@ -3,4 +3,4 @@ export * from './constants';
 export * from './Logger';
 export * from './msg';
 export type * from './types';
-export * from './webViewUtils';
+export * from './webViewContentUtils';
